+++ src/ResourceManager/DataMigration/Commands.DataMigration/Cmdlets/NewConnectionInfoCmdlet.cs
@@ -1,19 +1,27 @@
-﻿// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="NewConnectionInfoCmdlet.cs" company="Microsoft">
-//     Copyright (c) Microsoft Corporation.
-// </copyright>
-// --------------------------------------------------------------------------------------------------------------------
+﻿// ----------------------------------------------------------------------------------
+//
+// Copyright Microsoft Corporation
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+// http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// ----------------------------------------------------------------------------------
 
-using Microsoft.Azure.Management.DataMigration.Models;
 using System.Management.Automation;
 using Microsoft.Azure.Commands.DataMigration.Models;
+using Microsoft.Azure.Management.DataMigration.Models;
 
 namespace Microsoft.Azure.Commands.DataMigration.Cmdlets
 {
     /// <summary>
     /// Class that creates a new instance of the Sql Server Connection Info.
     /// </summary>
-    [Cmdlet(VerbsCommon.New, "AzureRmDataMigrationConnectionInfo", SupportsShouldProcess = true), OutputType(typeof(ConnectionInfo))]
+    [Cmdlet(VerbsCommon.New, "AzureRmDataMigrationConnectionInfo"), OutputType(typeof(ConnectionInfo))]
     [Alias("New-AzureRmDmsConnInfo")]
     public class NewConnectionInfoCmdlet : DataMigrationCmdlet, IDynamicParameters
     {
@@ -42,18 +50,15 @@ namespace Microsoft.Azure.Commands.DataMigration.Cmdlets
 
         public override void ExecuteCmdlet()
         {
-            if (ShouldProcess(this.ServerType.ToString(), Resources.createDbInfo))
-            {
-                base.ExecuteCmdlet();
+            base.ExecuteCmdlet();
 
-                if (connCmdlet != null)
-                {
-                    WriteObject(connCmdlet.ProcessConnectionInfoCmdlet());
-                }
-                else
-                {
-                    throw new PSArgumentException("Invalid Argument List");
-                }
+            if (connCmdlet != null)
+            {
+                WriteObject(connCmdlet.ProcessConnectionInfoCmdlet());
+            }
+            else
+            {
+                throw new PSArgumentException("Invalid Argument List");
             }
         }