+++ src/ResourceManager/Resources/Commands.Resources.Test/ScenarioTests/RoleAssignmentTests.cs
@@ -19,12 +19,12 @@ using Microsoft.Azure.Management.Authorization;
 using Microsoft.Azure.Management.ResourceManager;
 using Microsoft.Azure.Management.ResourceManager.Models;
 using Microsoft.Azure.ServiceManagemenet.Common.Models;
-using Microsoft.Azure.Test;
 using Microsoft.WindowsAzure.Commands.ScenarioTest;
 using Microsoft.WindowsAzure.Commands.Test.Utilities.Common;
 using Microsoft.WindowsAzure.Commands.Utilities.Common;
-using System;
 using System.Linq;
+using System.Reflection;
+using Microsoft.Rest.ClientRuntime.Azure.TestFramework;
 using Xunit;
 using Xunit.Abstractions;
 
@@ -50,8 +50,13 @@ namespace Microsoft.Azure.Commands.Resources.Test.ScenarioTests
         {
             ResourcesController.NewInstance.RunPsTest("Test-RaDeletedPrincipals");
         }
-	
+
+#if NETSTANDARD
+        [Fact(Skip = "DisableTestParallelization disabled on .NET Core: Test uses RoleDefinitionNames")]
+        [Trait(Category.RunType, Category.DesktopOnly)]
+#else
         [Fact]
+#endif
         [Trait(Category.AcceptanceType, Category.Flaky)]
         public void RaPropertiesValidation() {
             ResourcesController.NewInstance.RunPsTest("Test-RaPropertiesValidation");
@@ -70,7 +75,7 @@ namespace Microsoft.Azure.Commands.Resources.Test.ScenarioTests
         {
             ResourcesController.NewInstance.RunPsTest("Test-RaByScope");
         }
-        
+
         [Fact]
         [Trait(Category.AcceptanceType, Category.CheckIn)]
         public void RaDeleteByPSRoleAssignment()
@@ -123,14 +128,14 @@ namespace Microsoft.Azure.Commands.Resources.Test.ScenarioTests
             ResourcesController.NewInstance.RunPsTest("Test-RaDeletionByScope");
         }
 
-        [Fact]
+        [Fact(Skip = "Need AD team to re-record")]
         [Trait(Category.AcceptanceType, Category.CheckIn)]
         public void RaDeletionByScopeAtRootScope()
         {
             ResourcesController.NewInstance.RunPsTest("Test-RaDeletionByScopeAtRootScope");
         }
-        
-        [Fact]
+
+        [Fact(Skip = "Need AD team to re-record")]
         [Trait(Category.AcceptanceType, Category.CheckIn)]
         public void RaDelegation()
         {
@@ -217,12 +222,10 @@ namespace Microsoft.Azure.Commands.Resources.Test.ScenarioTests
                                 resourceGroup.Name)
                         };
                     },
-                    // initialize
+                    // cleanup
                     null,
-                    // cleanup 
-                    null,
-                    TestUtilities.GetCallingClass(),
-                    TestUtilities.GetCurrentMethodName() + "_Setup");
+                    MethodBase.GetCurrentMethod().ReflectedType?.ToString(),
+                    MethodBase.GetCurrentMethod().Name + "_Setup");
 
                 // login as different user and run the test
                 var controllerUser = ResourcesController.NewInstance;
@@ -241,19 +244,10 @@ namespace Microsoft.Azure.Commands.Resources.Test.ScenarioTests
                             userPermission)
                         };
                     },
-                    // initialize
-                    (testFactory) =>
-                    {
-                        if (newUser != null)
-                        {
-                            testFactory.CustomEnvValues[TestEnvironment.UserIdKey] = userName + "@" + controllerAdmin.UserDomain;
-                            testFactory.CustomEnvValues[TestEnvironment.AADPasswordKey] = userPass;
-                        }
-                    },
-                    // cleanup 
+                    // cleanup
                     null,
-                    TestUtilities.GetCallingClass(),
-                    TestUtilities.GetCurrentMethodName() + "_Test");
+                    MethodBase.GetCurrentMethod().ReflectedType?.ToString(),
+                    MethodBase.GetCurrentMethod().Name + "_Test");
             }
             finally
             {
@@ -264,7 +258,7 @@ namespace Microsoft.Azure.Commands.Resources.Test.ScenarioTests
                     null,
                     // initialize
                     null,
-                    // cleanup 
+                    // cleanup
                     () =>
                     {
                         if (newUser != null)
@@ -275,11 +269,11 @@ namespace Microsoft.Azure.Commands.Resources.Test.ScenarioTests
                         if (resourceGroup != null)
                         {
                             controllerAdmin.AuthorizationManagementClient.RoleAssignments.Delete(resourceGroup.Id, roleAssignmentId).ToString();
-                        }                        
+                        }
                     },
-                    TestUtilities.GetCallingClass(),
-                    TestUtilities.GetCurrentMethodName() + "_Cleanup");
-            }                       
+                    MethodBase.GetCurrentMethod().ReflectedType?.ToString(),
+                    MethodBase.GetCurrentMethod().Name + "_Cleanup");
+            }
         }
     }
 }