+++ src/ResourceManager/ManagementPartner/ChangeLog.md
@@ -19,10 +19,12 @@
 -->
 ## Current Release
 
+## Version 0.1.2
+* Set minimum dependency of module to PowerShell 5.0
+
 ## Version 0.1.1
 * Add Partner Name in the response for cmdlets: Get-AzureRmManagementPartner, New-AzureRmManagementPartner and Update-AzureRmManagementPartner
 * Updated to the latest version of the Azure ClientRuntime
-* Set minimum dependency of module to PowerShell 5.0
 
 ## Version 0.1.0
 * First release of Management Partners cmdlets