+++ src/ResourceManager/TrafficManager/ChangeLog.md
@@ -18,6 +18,11 @@
         - Additional information about change #1
 -->
 ## Current Release
+
+## Version 4.0.5
+* Update the parameters for `Get-AzureRmTrafficManagerProfile` so that -ResourceGroupName parameter is required when using -Name parameter.
+
+## Version 4.0.4
 * Set minimum dependency of module to PowerShell 5.0
 
 ## Version 4.0.3